train45 assembly fixup: NativeIntConstMask re-baselined at the union (seat 3's defect-D rule reaching an explicit-cast site)
Seat 3 (claude/g-root1-tag-and-cast 31668f43e, defect D: nativeWidthUnsignedPrefix applied on the
signed-parse branch of an above-MaxUint32 literal resolved to a native-width unsigned type) reaches
this guard's `(uintptr)0x00ffffffffffUL` operand, which sits inside an explicit cast and therefore
compiled before the fix; the zero-footprint argument covered assignment sites only.  The union
converter emits `(uintptr)(nuint)0x00ffffffffffUL` -- the rule's intended emission, measured by LEG 4
and LEG 4b (CNR at both pins, CHANGED 1, this project only), by i9's CNR at both pins on the same
tree, and by i9's forced transpile (numstat 1/1, this line).  Re-baselined with the union converter
at the two-pin pairing: main.cs and main.cs.target, numstat 1/1 each, golden byte-equal to the
emission.  The seat is unchanged at 31668f43e; this is an assembly commit, since no lane owns a golden.

+++ src/tests/Behavioral/NativeIntConstMask/main.cs
@@ -17,7 +17,7 @@ internal static uintptr align(uintptr i) {
 }
 
 internal static uintptr maskAddr(uintptr i) {
-    return (uintptr)(i & (uintptr)0x00ffffffffffUL);
+    return (uintptr)(i & (uintptr)(nuint)0x00ffffffffffUL);
 }
 
 internal static uintptr alignSmall(uintptr i) {
